gejmboj_cpu

Module bus

Bus implementation

The bus contains memory data which is stored in Little-Endian fashion, which means that the least significant byte is stored at a lower memory location than the most significant byte.

Memory map

0000-7FFF: External bus (ROM)
8000-9FFF: VRAM
A000-BFFF: External bus (RAM)
C000-DFFF: WRAM
E000-FDFF: Echo (WRAM)
FE00-FE9F: Object Attribute Memory (OAM)
FEA0-FEFF: Invalid OAM
FF00-FF7F: Memory mapped I/O
FF80-FFFE: High RAM (HRAM)
FFFF:      IE register

Structs

Bus

Enums

Clock

Constants

ADDR_BGP

ADDR_DIV

ADDR_IE

ADDR_IF

ADDR_LCDC

ADDR_LYC

ADDR_NR10

ADDR_NR11

ADDR_NR12

ADDR_NR14

ADDR_NR21

ADDR_NR22

ADDR_NR24

ADDR_NR30

ADDR_NR31

ADDR_NR32

ADDR_NR34

ADDR_NR41

ADDR_NR42

ADDR_NR43

ADDR_NR44

ADDR_NR50

ADDR_NR51

ADDR_NR52

ADDR_OBP0

ADDR_OBP1

ADDR_SCX

ADDR_SCY

ADDR_TAC

ADDR_TIMA

ADDR_TMA

ADDR_WX

ADDR_WY

MASK_TIMER_ENABLED