Module gejmboj_cpu::instructions

Sharp SM83 instruction set

Structs

Call
Unconditional call of the function at operand address.

CallIf
Conditional function call.

Jp
Unconditional jump to location specified by 16-bit operand.

JpIf
Conditional jump to location specified by 16-bit operand.

JpToHL
Unconditional jump to location specified by register HL

JpToOffset
Unconditional jump to location at current + offset

JpToOffsetIf
Coonditional jump to relative address specified by offset operand.

Noop
No operation

Ret
Unconditional return from function.

RetIf
Conditionally return from function.

Enums

Condition

Traits

Instruction
Trait for implementing a Sharp SM83 instruction.

Functions

decode
Decode an operation code into an Instruction.

Type Definitions

InstructionResult
Return either the number of consumed machine cycles, or a CpuError.