gejmboj_cpu::instructions::alu_8bit

Enum ALU8Bit

pub enum ALU8Bit {

    ADD(SingleRegister),
    ADD_N(u8),
    ADD_HL(),
    ADC(SingleRegister),
    ADC_N(u8),
    ADC_HL(),
    SUB(SingleRegister),
    SUB_N(u8),
    SUB_HL(),
    SBC(SingleRegister),
    SBC_N(u8),
    SBC_HL(),
    AND(SingleRegister),
    AND_N(u8),
    AND_HL(),
    OR(SingleRegister),
    OR_N(u8),
    OR_HL(),
    XOR(SingleRegister),
    XOR_N(u8),
    XOR_HL(),
    CP(SingleRegister),
    CP_N(u8),
    CP_HL(),
    INC(SingleRegister),
    INC_HL(),
    DEC(SingleRegister),
    DEC_HL(),

}

8-bit ALU (math) instructions

Variants

ADD(SingleRegister)

Add value of SingleRegister to A

ADD_N(u8)

Add value of operand to A

ADD_HL()

Add value of (HL) to A

ADC(SingleRegister)

Add value of SingleRegister and the Carry flag to A

ADC_N(u8)

Add value of operand and Carry to A

ADC_HL()

Add value of (HL) and Carry to A

SUB(SingleRegister)

Subtract value of SingleRegister from A

SUB_N(u8)

Subtract value of operand from A

SUB_HL()

Subtract value of (HL) from A

SBC(SingleRegister)

Subtract value of SingleRegister and Carry from A

SBC_N(u8)

Subtract value of operand and Carry from A

SBC_HL()

Subtract value of (HL) and Carry from A

AND(SingleRegister)

Logical AND between register and A

AND_N(u8)

Logical AND between operand and A

AND_HL()

Logical AND between (HL) and A

OR(SingleRegister)

Logical OR between register and A

OR_N(u8)

Logical OR between operand and A

OR_HL()

Logical OR between (HL) and A

XOR(SingleRegister)

Logical XOR between register and A

XOR_N(u8)

Logical XOR between operand and A

XOR_HL()

Logical XOR between (HL) and A

CP(SingleRegister)

Compare register and A

Basically an A - n subtraction but with the result being thrown away, so the same flag rules as Sub apply.

CP_N(u8)

Compare operand and A

CP_HL()

Compare (HL) and A

INC(SingleRegister)

Increment SingleRegister by 1

The Carry flag is unaffected by this instruction.

INC_HL()

Increment the address pointed to by HL by 1

The Carry flag is unaffected by this instruction.

DEC(SingleRegister)

Decrement SingleRegister by 1

The Carry flag is unaffected by this instruction.

DEC_HL()

Decrement HL by 1

The Carry flag is unaffected by this instruction.

Implementations

impl ALU8Bit

pub fn execute( &self, registers: &mut Registers, memory: &mut Bus, _cpu_flags: &mut CpuFlags, ) -> InstructionResult

pub const fn length(&self) -> u16