Enum gejmboj_cpu::instructions::load_8bit::Load8Bit

pub enum Load8Bit {

    LD(SingleRegister, SingleRegister),
    LD_FROM_HL(SingleRegister),
    LD_TO_HL(SingleRegister),
    LD_N(SingleRegister, u8),
    LD_N_TO_HL(u8),
    LD_BC_TO_A(),
    LD_DE_TO_A(),
    LD_A_TO_BC(),
    LD_A_TO_DE(),
    LD_TO_A(u16),
    LD_FROM_A(u16),
    LDH_C_TO_A(),
    LDH_C_FROM_A(),
    LDH_TO_A(u8),
    LDH_FROM_A(u8),
    LD_A_FROM_HL_DEC(),
    LD_A_TO_HL_DEC(),
    LD_A_FROM_HL_INC(),
    LD_A_TO_HL_INC(),

}

8 Bit load instructions.

Variants

LD(SingleRegister, SingleRegister)

Loads data from register r2 into r1.

LD_FROM_HL(SingleRegister)

Loads data pointed to by HL into r.

LD_TO_HL(SingleRegister)

Loads data in r into location pointed to by HL.

LD_N(SingleRegister, u8)

Loads operand into register r.

LD_N_TO_HL(u8)

Load the value of operand into the location pointed to by HL

LD_BC_TO_A()

Load data at address pointed to by BC into A

LD_DE_TO_A()

Load data at address pointed to by DE into A

LD_A_TO_BC()

Load A into into address pointed to by BC

LD_A_TO_DE()

Load A into into address pointed to by DE

LD_TO_A(u16)

Load data at address into A

LD_FROM_A(u16)

Load data in A into address at address

LDH_C_TO_A()

Load data to A from the address at 0xFF00 + register C

LDH_C_FROM_A()

Load data from A into the address at 0xFF00 + register C

LDH_TO_A(u8)

Load data to A from the address at 0xFF00 + operand

LDH_FROM_A(u8)

Load data from A into the address at 0xFF00 + operand

LD_A_FROM_HL_DEC()

Load data to A from the address at HL, value at HL is decremented.

LD_A_TO_HL_DEC()

Load data to address at HL from A, HL is decremented after write.

LD_A_FROM_HL_INC()

Load data to A from the address at HL, value at HL is incremented.

LD_A_TO_HL_INC()

Load data to address at HL from A, HL is incremented after write.

Implementations

impl Load8Bit

pub fn execute( &self, registers: &mut Registers, memory: &mut Memory, _cpu_flags: &mut CpuFlags ) -> InstructionResult

pub fn length(&self) -> u16