Enum gejmboj_cpu::instructions::misc::Misc

pub enum Misc {
    NOP(),
    DI(),
    EI(),
    CCF(),
    SCF(),
    DAA(),
    CPL(),
}

Miscelleneous instructions

Some instructions operate on the flag register F:

,---.---.---.---.---.---.---.---.
| 7 | 6 | 5 | 4 | 3 | 2 | 1 | 0 | Bit
|---|---|---|---|---|---|---|---|
| Z | N | H | C | 0 | 0 | 0 | 0 | Flag
`---´---´---´---´---´---´---´---´

Where bit 0-3 are grounded to 0 and can’t be overwritten and C is for carry, H for half-carry, N for negative and Z for zero.

Variants

NOP()

No operation

DI()

Disable interrupt handling

EI()

Schedules interrupt handling to be enabled after the next machine cycle

CCF()

Flips the carry flag (C) and clears the negative (N) and half-carry (H) flags

SCF()

Sets the carry flag (C) and clears the negative (N) and half-carry (H) flags

DAA()

Decimal Adjust Accumulator (DAA)

This instruction affects the A register and should be called after Binary Coded Decimal (BCD) addition or subtraction instructions. This instruction converts the value stored in A into BCD representation. The instruction sets the Carry and Zero flags if appropriate.

In BCD representation each nibble (4 bits) represents a digit:

Decimal	Binary	BCD

1	0b0000_0001	0b0000_0001 (0, 1)

10	0b0000_1010	0b0001_0000 (1, 0)

28	0b0001_1100	0b0010_1000 (2, 8)

When you operate on BCD numbers you have to convert the result into BCD as well according to the following rules:

Add 6 to each digit above 9 if addition

Subtract 6 from each digit above 9 if subtraction

All subtraction is achieved by adding the Two’s Complement (inverse of number + 1).

BCD example: 29 + 13 = 42

  0b0010_1001
+ 0b0001_0011
  -----------
  0b0011_1100 (BCD: 3(12), Binary: 60)

  0b0011_1100
+ 0b0000_0110
  -----------
  0b0100_0010 (BCD: 42, Binary: 66)

BCD example: 23 - 19 = 4

    11     11
  0b0010_0011
+ 0b1110_0111 (Two's Complement of 19)
  -----------
  0b0000_1010 (BCD: 8, Binary: 8)

    1111   1
  0b0000_1010
+ 0b1111_1010
  -----------
  0b0000_0100 (BCD: 4, Binary: 4)

BCD representation is often used instead of converting back-and-forth between binary and decimal when doing addition and subtraction, especially when no micro-processor is involved since the necessary circuit becomes a lot simpler. A common use-case is Seven Segment Displays where each display represents a digit.

CPL()

Flips all bits in the A register and sets the negative (N) and half-carry (H) flags

Implementations

impl Misc

pub fn execute( &self, registers: &mut Registers, _memory: &mut Memory, cpu_flags: &mut CpuFlags ) -> InstructionResult

pub fn length(&self) -> u16