Enum gejmboj_cpu::instructions::rotate_shift::RotateShift

pub enum RotateShift {
    RLCA(),
    RLA(),
    RRCA(),
    RRA(),
    RLC(u8),
    RL(u8),
    RRC(u8),
    RR(u8),
    SLA(u8),
    SRA(u8),
    SRL(u8),
    SWAP(u8),
}

Bit rotate and shift instructions.

Some instructions operate on m which is indicated by the 8-bit operand passed to the instruction as per the following table:

Operand	Target

0000_0110	(HL), the memory contents pointed to by the HL register

0000_0rrr	8-bit register rrr

Variants

RLCA()

Rotate contents of register A to the left. Bit 7 is placed in both C and Bit 0.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A7

RLA()

Rotates contents of register A to the left. C is put in A0 and A7 is put in C.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A7

RRCA()

Rotate contents of register A to the right. Bit 0 is placed in both C and Bit 7.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A0

RRA()

Rotates contents of register A to the right. C is put in A7 and A0 is put in C.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A0

RLC(u8)

Rotates contents of m to the left.

m7 is copied to both C and m0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

RL(u8)

Rotates contents of m to the left.

C is copied to m0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

RRC(u8)

Rotates contents of m to the right.

m0 is copied to both C and m7.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

RR(u8)

Rotates contents of m to the right.

C is copied to m7.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

SLA(u8)

Shifts the contents of m to the left.

m7 is copied to C and m0 is reset to 0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

SRA(u8)

Shifts the contents of m to the right.

m7 is unchanged and m0 is copied to C.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

SRL(u8)

Shifts the contents of m to the right.

m7 is set to 0 and m0 is copied to C.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

SWAP(u8)

Swaps the high and low nibble of m.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	0

Implementations

impl RotateShift

pub fn execute( &self, registers: &mut Registers, memory: &mut Memory, _cpu_flags: &mut CpuFlags ) -> InstructionResult

pub fn length(&self) -> u16