Enum gejmboj_cpu::instructions::alu_8bit::ALU8Bit

pub enum ALU8Bit {

    Add(SingleRegister),
    AddN(u8),
    AddHL(),
    Adc(SingleRegister),
    AdcN(u8),
    AdcHL(),
    Sub(SingleRegister),
    SubN(u8),
    SubHL(),
    Sbc(SingleRegister),
    SbcN(u8),
    SbcHL(),
    And(SingleRegister),
    AndN(u8),
    AndHL(),
    Or(SingleRegister),
    OrN(u8),
    OrHL(),
    Xor(SingleRegister),
    XorN(u8),
    XorHL(),
    Cp(SingleRegister),
    CpN(u8),
    CpHL(),
    Inc(SingleRegister),
    IncHL(),
    Dec(SingleRegister),
    DecHL(),

}

8-bit ALU (math) instructions

Variants

Add(SingleRegister)

Add value of SingleRegister to A

Tuple Fields of Add

0: SingleRegister

AddN(u8)

Add value of operand to A

Tuple Fields of AddN

0: u8

AddHL()

Add value of (HL) to A

Tuple Fields of AddHL

Adc(SingleRegister)

Add value of SingleRegister and the Carry flag to A

Tuple Fields of Adc

0: SingleRegister

AdcN(u8)

Add value of operand and Carry to A

Tuple Fields of AdcN

0: u8

AdcHL()

Add value of (HL) and Carry to A

Tuple Fields of AdcHL

Sub(SingleRegister)

Subtract value of SingleRegister from A

Tuple Fields of Sub

0: SingleRegister

SubN(u8)

Subtract value of operand from A

Tuple Fields of SubN

0: u8

SubHL()

Subtract value of (HL) from A

Tuple Fields of SubHL

Sbc(SingleRegister)

Subtract value of SingleRegister and Carry from A

Tuple Fields of Sbc

0: SingleRegister

SbcN(u8)

Subtract value of operand and Carry from A

Tuple Fields of SbcN

0: u8

SbcHL()

Subtract value of (HL) and Carry from A

Tuple Fields of SbcHL

And(SingleRegister)

Logical AND between register and A

Tuple Fields of And

0: SingleRegister

AndN(u8)

Logical AND between operand and A

Tuple Fields of AndN

0: u8

AndHL()

Logical AND between (HL) and A

Tuple Fields of AndHL

Or(SingleRegister)

Logical OR between register and A

Tuple Fields of Or

0: SingleRegister

OrN(u8)

Logical OR between operand and A

Tuple Fields of OrN

0: u8

OrHL()

Logical OR between (HL) and A

Tuple Fields of OrHL

Xor(SingleRegister)

Logical XOR between register and A

Tuple Fields of Xor

0: SingleRegister

XorN(u8)

Logical XOR between operand and A

Tuple Fields of XorN

0: u8

XorHL()

Logical XOR between (HL) and A

Tuple Fields of XorHL

Cp(SingleRegister)

Compare register and A

Basically an A - n subtraction but with the result being thrown away, so the same flag rules as Sub apply.

Tuple Fields of Cp

0: SingleRegister

CpN(u8)

Compare operand and A

Tuple Fields of CpN

0: u8

CpHL()

Compare (HL) and A

Tuple Fields of CpHL

Inc(SingleRegister)

Increment SingleRegister by 1

Tuple Fields of Inc

0: SingleRegister

IncHL()

Increment HL by 1

Tuple Fields of IncHL

Dec(SingleRegister)

Decrement SingleRegister by 1

Tuple Fields of Dec

0: SingleRegister

DecHL()

Decrement HL by 1

Tuple Fields of DecHL

Implementations

impl ALU8Bit

pub fn execute(
&self,
registers: &mut Registers,
memory: &mut Memory,
_cpu_flags: &mut CpuFlags
) -> InstructionResult

pub fn length(&self) -> u16