Enum gejmboj_cpu::instructions::rotate_shift::RotateShift

pub enum RotateShift {
    RlcA(),
    RlA(),
    RrcA(),
    RrA(),
    RlC(u8),
    Rl(u8),
    RrC(u8),
    Rr(u8),
    Sla(u8),
    Sra(u8),
    Srl(u8),
    Swap(u8),
}

Bit rotate and shift instructions.

Some instructions operate on m which is indicated by the 8-bit operand passed to the instruction as per the following table:

Operand	Target

0000_0110	(HL), the memory contents pointed to by the HL register

0000_0rrr	8-bit register rrr

Variants

RlcA()

Rotate contents of register A to the left. Bit 7 is placed in both C and Bit 0.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A7

Tuple Fields of RlcA

RlA()

Rotates contents of register A to the left. C is put in A0 and A7 is put in C.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A7

Tuple Fields of RlA

RrcA()

Rotate contents of register A to the right. Bit 0 is placed in both C and Bit 7.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A0

Tuple Fields of RrcA

RrA()

Rotates contents of register A to the right. C is put in A7 and A0 is put in C.

Flags

Flag	Effect

Z	0

N	0

H	0

C	A0

Tuple Fields of RrA

RlC(u8)

Rotates contents of m to the left.

m7 is copied to both C and m0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

Tuple Fields of RlC

0: u8

Rl(u8)

Rotates contents of m to the left.

C is copied to m0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

Tuple Fields of Rl

0: u8

RrC(u8)

Rotates contents of m to the right.

m0 is copied to both C and m7.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

Tuple Fields of RrC

0: u8

Rr(u8)

Rotates contents of m to the right.

C is copied to m7.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

Tuple Fields of Rr

0: u8

Sla(u8)

Shifts the contents of m to the left.

m7 is copied to C and m0 is reset to 0.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m7

Tuple Fields of Sla

0: u8

Sra(u8)

Shifts the contents of m to the right.

m7 is unchanged and m0 is copied to C.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

Tuple Fields of Sra

0: u8

Srl(u8)

Shifts the contents of m to the right.

m7 is set to 0 and m0 is copied to C.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	m0

Tuple Fields of Srl

0: u8

Swap(u8)

Swaps the high and low nibble of m.

Flags

Flag	Effect

Z	Set if 0

N	0

H	0

C	0

Tuple Fields of Swap

0: u8

Implementations

impl RotateShift

pub fn execute(
&self,
registers: &mut Registers,
memory: &mut Memory,
_cpu_flags: &mut CpuFlags
) -> InstructionResult

pub fn length(&self) -> u16