Module gejmboj_cpu::registers

Sharp SM83 register

16-bit special purpose registers

PC (Program counter)

SP (Stack pointer)

General purpose registers

,---.---.
| A | F |
|---|---|
| B | C |
|---|---|
| D | E |
|---|---|
| H | L |
`---´---´

Each register is 8-bit, but can be combined for 16-bit values according to the table rows above.

Flag register F

Register F is the Flag register with the following flag mapping:

,---.---.---.---.---.---.---.---.
| 7 | 6 | 5 | 4 | 3 | 2 | 1 | 0 | Bit
|---|---|---|---|---|---|---|---|
| Z | N | H | C | 0 | 0 | 0 | 0 | Flag
`---´---´---´---´---´---´---´---´

Flag	Name	Description

Z	Zero	Set if result of math operation was 0 or if two values match after CP

N	Negative	Set if last math operation was a subtraction

H	Half-carry	Set if carry occurred from the lower nibble (4 bits) in the last math operation

C	Carry	Set if carry occurred last math operation or if A is the smaller value after CP

Bits 0-3 are grounded to 0 and can’t be overwritten.

Stack pointer register (SP)

The stack pointer register is initialized to 0xFFFE and grows top-down, which means it is decremented.

Structs

Registers

Enums

DoubleRegister

Represents a 16-bit general purpose register.

SingleRegister

Represents an 8-bit general purpose register.

Constants

MASK_FLAG_CARRY

MASK_FLAG_HALF_CARRY

MASK_FLAG_NEGATIVE

MASK_FLAG_ZERO