Module gejmboj_cpu::registers

Sharp SM83 register

16-bit special purpose registers

PC (Program counter)

SP (Stack pointer)

General purpose registers

,---.---.
| A | F |
|---|---|
| B | C |
|---|---|
| D | E |
|---|---|
| H | L |
`---´---´

Each register is 8-bit, but can be combined for 16-bit values according to the table rows above.

Flag register F

Register F is the Flag register with the following flag mapping:

,---.---.---.---.---.---.---.---.
| 7 | 6 | 5 | 4 | 3 | 2 | 1 | 0 | Bit
|---|---|---|---|---|---|---|---|
| Z | N | H | C | 0 | 0 | 0 | 0 | Flag
`---´---´---´---´---´---´---´---´

Where bit 0-3 are grounded to 0 and can’t be overwritten and C is for carry, H for half-carry, N for negative and Z for zero.

Stack pointer register (SP)

The stack pointer register is initialized to 0xFFFE and grows top-down, which means it is decremented.

Structs

Registers

Enums

DoubleRegister

Represents a 16-bit general purpose register.

SingleRegister

Represents an 8-bit general purpose register.