Struct gejmboj_cpu::registers::Registers

pub struct Registers {
    pub PC: u16,
    pub SP: u16,
    /* private fields */
}

Fields
PC: u16SP: u16
Implementations

impl Registers

pub fn new() -> Self

pub fn set_single(&mut self, r: &SingleRegister, value: u8)

Sets the value of a SingleRegister.

Examples

let mut registers = Registers::new();

registers.set_single(&SingleRegister::A, 42);

assert_eq!(42, registers.get_single(&SingleRegister::A));

Special cases

Lowest nibble of register F is always 0 and can’t be overwritten.

registers.set_single(&SingleRegister::F, 0xFF);

assert_eq!(0xF0, registers.get_single(&SingleRegister::F));

pub fn get_single(&self, r: &SingleRegister) -> u8

Gets the current value of a SingleRegister.

Examples

let registers = Registers::new();

assert_eq!(0, registers.get_single(&SingleRegister::A));

pub fn get_double(&self, r: &DoubleRegister) -> u16

Gets value from a double 16-bit register

Examples

let mut registers = Registers::new();
registers.set_single(&SingleRegister::B, 0xAB);
registers.set_single(&SingleRegister::C, 0xCD);

assert_eq!(0xABCD, registers.get_double(&DoubleRegister::BC));

pub fn set_double(&mut self, r: &DoubleRegister, value: u16)

Sets value of a double 16-bit register

Examples

let mut registers = Registers::new();
registers.set_double(&DoubleRegister::BC, 0xAABB);

assert_eq!(0xAABB, registers.get_double(&DoubleRegister::BC));

Special cases

Lowest nibble of DoubleRegister::AF is always 0 and cannot be overwritten.

let mut registers = Registers::new();
registers.set_double(&DoubleRegister::AF, 0xABCD);

assert_eq!(0xABC0, registers.get_double(&DoubleRegister::AF));

pub fn increment_sp(&mut self) -> u16

Increments the SP by 2 and returns new SP value.

Examples

let mut registers = Registers::new();

assert_eq!(0xFFFE, registers.get_double(&DoubleRegister::SP));

registers.decrement_sp();
registers.increment_sp();
assert_eq!(0xFFFE, registers.get_double(&DoubleRegister::SP));

pub fn decrement_sp(&mut self) -> u16

Decrements the SP by 2 and returns new SP value.

Examples

let mut registers = Registers::new();

assert_eq!(0xFFFE, registers.get_double(&DoubleRegister::SP));

registers.decrement_sp();
assert_eq!(0xFFFC, registers.get_double(&DoubleRegister::SP));

pub fn is_carry(&self) -> bool

Returns true if the carry flag is set.

Examples

let mut registers = Registers::new();

registers.set_flags(0b0000_0000);
assert_eq!(false, registers.is_carry());

registers.set_flags(0b0001_0000);
assert_eq!(true, registers.is_carry());

pub fn is_half_carry(&self) -> bool

Returns true if the half carry flag is set.

Examples

let mut registers = Registers::new();

registers.set_flags(0b0000_0000);
assert_eq!(false, registers.is_half_carry());

registers.set_flags(0b0010_0000);
assert_eq!(true, registers.is_half_carry());

pub fn is_negative(&self) -> bool

Returns true if the negative flag is set.

Examples

let mut registers = Registers::new();

registers.set_flags(0b0000_0000);
assert_eq!(false, registers.is_negative());

registers.set_flags(0b0100_0000);
assert_eq!(true, registers.is_negative());

pub fn is_zero(&self) -> bool

Returns true if the zero flag is set.

Examples

let mut registers = Registers::new();

registers.set_flags(0b0000_0000);
assert_eq!(false, registers.is_zero());

registers.set_flags(0b1000_0000);
assert_eq!(true, registers.is_zero());

pub fn get_flags(&self) -> u8

Returns the value of the flag register F.

Convenience method to get the flag register value.

pub fn set_flags(&mut self, flags: u8)

Sets the value of the flag register F.

Convenience method to set the flag register value. The lower nibble (4 bits) are always 0 and cannot be overwritten.

Examples

let mut registers = Registers::new();

registers.set_flags(0b1101_0000);
assert_eq!(0b1101_0000, registers.get_flags());

registers.set_flags(0b1111_1111);
assert_eq!(0b1111_0000, registers.get_flags());

pub fn set_carry(&mut self, set: bool)

Convenience function to set or reset the carry flag.

Examples

let mut registers = Registers::new();

registers.set_carry(true);
assert_eq!(true, registers.is_carry());

registers.set_carry(false);
assert_eq!(false, registers.is_carry());

pub fn set_half_carry(&mut self, set: bool)

Convenience function to set or reset the half-carry flag.

Examples

let mut registers = Registers::new();

registers.set_half_carry(true);
assert_eq!(true, registers.is_half_carry());

registers.set_half_carry(false);
assert_eq!(false, registers.is_half_carry());

pub fn set_negative(&mut self, set: bool)

Convenience function to set or reset the negative flag.

Examples

let mut registers = Registers::new();

registers.set_negative(true);
assert_eq!(true, registers.is_negative());

registers.set_negative(false);
assert_eq!(false, registers.is_negative());

pub fn set_zero(&mut self, set: bool)

Convenience function to set or reset the zero flag.

Examples

let mut registers = Registers::new();

registers.set_zero(true);
assert_eq!(true, registers.is_zero());

registers.set_zero(false);
assert_eq!(false, registers.is_zero());